iot_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
redis
kafka-python
python-dotenv
cryptography